Use carriage return over escape sequence
Before, when clearing a line, the cursor would be moved left by a very
large number in order to go to the beginning of the line. This was a
stupid way to do it when carriage return `\r` goes to the beginning of
the line and it will always work, regardless of how large the input line
was.