Replace stream locks with generics in Environment
Environment now takes in 2 generics, Read and Write (for now). Before,
it used to require an `StdinLock` and an `StdoutLock`. With generics it
enables the ability to work with files in the future (as explained in
previous commit). A side effect is that now the effective `main` of the
program is actually `src/config.rs::initialise()` since main doesn't
accept type generics.