Some basic sh parsing.
Now every single command is tokenized and then processed. This is much
more advanced than the previous parsing which was just based on
splitting on spaces and `|`. Also, This tokenizer respects quotes and
can even escape them with `\`.

This means that when previously `echo "hello world"` would output
`"hello world"`, with the new parser the output would be `hello world`
(which is correct).